перемещает все моковые данные в один файл
