+++ src/index.js
@@ -2,8 +2,7 @@ import React from 'react';
 import ReactDOM from 'react-dom';
 
 import App from './components/app/app.jsx';
-import {promoInfo} from './mock.js';
-import {moviesData} from './mocks/movies.js';
+import {moviesData, promoInfo} from './mocks/movies.js';
 
 const onMovieTitleClick = () => {};
